refactor: 긴 코드 방지를 위해 MemberTableRow 분리
